Refactor document<->program boundary
