+++ editor/language_server/src/program/services/highlight.rs
@@ -1,6 +1,6 @@
 //! Module `highlight` provides semantic highlight services.
 
-use super::local_response::LocalHighlight;
+use super::response::*;
 use crate::program::services::references as refs;
 use crate::Program;
 
@@ -10,13 +10,13 @@ impl Program {
     /// Returns semantic highlighting ranges relevant to a offset in a document.
     /// If the offset is over an identifier (variable), all references to that variable are returned.
     /// Otherwise, nothing is returned.
-    pub fn get_semantic_highlights(&self, offset: usize) -> Option<Vec<LocalHighlight>> {
+    pub fn get_semantic_highlights(&self, offset: usize) -> Option<Vec<ProgramHighlight>> {
         // The nice thing is that the references service already does most of the work to get
         // references, so we can just piggyback off that and translate types accordingly.
         let references = self.get_kinded_references(offset)?;
         let references = references
             .into_iter()
-            .map(|rk| LocalHighlight {
+            .map(|rk| ProgramHighlight {
                 span: *rk.span(),
                 kind: match rk {
                     refs::ReferenceKind::Definition(_) => DocumentHighlightKind::Write,
